requirements.txt
streamlit
openai-whisper
torch
torchaudio
torchvision
transformers
librosa
soundfile
numpy
pandas
plotly
speechbrain
langchain
openai
scikit-learn
